build: Upgraded target frameworks to net8.0 and net10.0 and updated NuGets
