+++ sandbox/Gui/Assets/avalonia-logo.ico
