+++ sandbox/Gui/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "GuiNet6": {
+    "Gui": {
       "commandName": "Project"
     },
     "WSL": {
