+++ sandbox/Gui/Views/DashboardView.axaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="400" d:DesignHeight="300"
-             x:Class="GuiNet6.Views.DashboardView">
+             x:Class="Gui.Views.DashboardView">
   <DockPanel>
     <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">
       <TextBlock Text="Welcome to Prism.Avalonia!" />
