+++ sandbox/Gui/Views/ShellWindow.axaml
@@ -2,12 +2,12 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:core="clr-namespace:GuiNet6;assembly=GuiNet6"
-        xmlns:local="using:GuiNet6.Views"
+        xmlns:core="clr-namespace:Gui;assembly=Gui"
+        xmlns:local="using:Gui.Views"
         xmlns:prism="http://prismlibrary.com/"
         prism:ViewModelLocator.AutoWireViewModel="True"
         mc:Ignorable="d" d:DesignWidth="600" d:DesignHeight="350"
-        x:Class="GuiNet6.Views.ShellWindow"
+        x:Class="Gui.Views.ShellWindow"
         Background="Transparent"
         Icon="/Assets/avalonia-logo.ico"
         Width="600" Height="300"
