+++ sandbox/Gui/Views/ShellWindow.axaml.cs
@@ -2,13 +2,16 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace GuiNet6.Views;
+namespace Gui.Views;
 
-public partial class SidebarView : UserControl
+public partial class ShellWindow : Window
 {
-  public SidebarView()
+  public ShellWindow()
   {
     InitializeComponent();
+#if DEBUG
+          this.AttachDevTools();
+#endif
   }
 
   private void InitializeComponent()
