+++ sandbox/Gui/Views/SidebarView.axaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="GuiNet6.Views.SidebarView">
+             x:Class="Gui.Views.SidebarView">
   <DockPanel>
 
     <TextBlock DockPanel.Dock="Top"
