--- sandbox/GuiNet6/nuget.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!-- 
-  To use the Avalonia CI feed to get unstable packages, move this file to the root of your solution.
--->
-
-<configuration>
-  <packageSources>
-    <add key="AvaloniaCI" value="https://www.myget.org/F/avalonia-ci/api/v2" />
-  </packageSources>
-</configuration>
