+++ src/LinuxDebugger.sln
@@ -13,9 +13,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{42E06A3F-6EF1-49EC-910E-A89C15BC6E91}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GuiNet6", "..\sandbox\GuiNet6\GuiNet6.csproj", "{2F01BAAF-DF1E-406C-A46D-127896C203F1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gui", "..\sandbox\Gui\Gui.csproj", "{2F01BAAF-DF1E-406C-A46D-127896C203F1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleNet6", "..\sandbox\ConsoleNet6\ConsoleNet6.csproj", "{D108A8EB-E4CD-4A84-A92B-DAD4DB6CA983}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Console", "..\sandbox\Console\Console.csproj", "{D108A8EB-E4CD-4A84-A92B-DAD4DB6CA983}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
