+++ src/VsLinuxDebugger/VsLinuxDebugger.csproj
@@ -95,16 +95,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.VisualStudio.SDK" Version="17.0.32112.339" ExcludeAssets="runtime" />
-    <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="17.2.1087-preview2" />
+    <PackageReference Include="Microsoft.VisualStudio.SDK" Version="17.14.40265" ExcludeAssets="runtime" />
+    <PackageReference Include="Microsoft.VSSDK.BuildTools" Version="17.12.2069" />
     <PackageReference Include="SharpCompress">
-      <Version>0.30.1</Version>
+      <Version>0.50.4</Version>
     </PackageReference>
     <PackageReference Include="SSH.NET">
       <Version>2023.0.1</Version>
     </PackageReference>
     <PackageReference Include="System.Text.Json">
-      <Version>6.0.10</Version>
+      <Version>10.0.11</Version>
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
